pyproject.toml
[tool.poetry]
name = "aidlp"
version = "1.9.0"
description = "AI Data Loss Prevention Proxy"
authors = ["Fabrizio Salmi <fabrizio.salmi@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
mitmproxy = "10.2.3"
flashtext = "2.7"
presidio-analyzer = "2.2.352"
presidio-anonymizer = "2.2.352"
typer = "0.9.0"
pyyaml = "6.0.1"
spacy = "3.7.2"
prometheus-client = "0.19.0"
python-json-logger = "2.0.7"
hvac = "2.1.0"
pybreaker = "^1.3.0"
pydantic = "^2.5.0"
requests = "^2.31.0"  # Needed for CLI stats fetching

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
locust = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
